refactor: PainLevel 뷰 리팩토링 및 UX 개선
- PainLevel 뷰의 코드를 정리하고 중복 파일을 삭제했습니다.
- 캘린더의 날짜 형식이 시스템 언어 설정에 맞게 표시되도록 수정했습니다.
- 더 명확한 의미 전달을 위해 영문 UX 라이팅을 간결하게 다듬었습니다.
